refactor inheritance to include person and instructor classes
