+++ apprentice_teacher.rb
@@ -1,21 +1,23 @@
-class ApprenticeTeacher
+require_relative 'instuctor'
+require_relative 'person'
+
+class ApprenticeTeacher < Instructor
   attr_reader :age, :salary, :phase, :target_raise
   attr_accessor :name
 
-  def initialize(options={})
-    @age = options.fetch(:age, 0)
-    @name = options.fetch(:name, "")
-    @target_raise = 800
-    @phase = 3
-  end
+  INITIAL_PHASE = 3
+  TARGET_RAISE = 800
 
-  def offer_high_five
-    "High five!"
-  end
+  # def initialize(options={})
+  #   @age = options.fetch(:age, 0)
+  #   @name = options.fetch(:name, "")
+  #   @target_raise = 800
+  #   @phase = 3
+  # end
 
-  def set_phase(num)
-    @phase = num
-    "Cool, I've always wanted to teach phase #{num}!"
+
+  def phase=(phase)
+    @phase = phase
   end
 
   def teach_stuff
@@ -26,20 +28,11 @@ class ApprenticeTeacher
     response
   end
 
-  def salary=(new_salary)
-    puts "This better be good!"
-    @salary = new_salary
-  end
-
-  def receive_raise(raise)
-    @salary += raise
-  end
-
   def set_performance_rating(rating)
     response = ""
     if rating > 80
       response = "Yay, I'm a great employee!"
-      receive_raise(@target_raise)
+      @target_raise = receive_raise(TARGET_RAISE)
     else
       response = "Oh, well -- thanks to this actionable, specific, and kind feedback, I'll do better next time."
     end
